Lift model-agnostic runtime types into kestrel.runtime
moondream/runtime.py was carrying types that other modules already treat
as cross-cutting (token kinds, sequence/prefill state). Move them into
a dedicated kestrel.runtime package so the engine and scheduler can
depend on a stable shared surface. moondream.runtime keeps re-exports
so all current importers (skills, scheduler, tests, scripts) stay
unchanged.

Moved into kestrel/runtime/tokens.py:
- TextToken, CoordToken, SizeToken, Token

Moved into kestrel/runtime/state.py:
- RuntimeDecodeResult, SequenceState, _CacheLookupResult,
  PrefillClassification, PreparedSequence

PrefillScratch and PrefillSlot stay in moondream.runtime because their
shapes are tied to Moondream's QKV/MoE config; they are not generic.
_LayerPagedCache and MoondreamRuntime stay as well.

No behavior change: 242 unit tests pass.
